Bundling and minification complete
JS reworked, concatenated and minified.
