+++ Project/App_Start/BundleConfig.cs
@@ -8,13 +8,9 @@ namespace Project
 		public static void RegisterBundles(BundleCollection bundles)
 		{
 			bundles.Add(new ScriptBundle("~/bundles/jquery").Include(
-						"~/Scripts/jquery-{version}.js"));
-
-			bundles.Add(new ScriptBundle("~/bundles/js.cookie").Include(
-				"~/Scripts/js.cookie.js"));
-
-			bundles.Add(new ScriptBundle("~/bundles/jqueryval").Include(
-						"~/Scripts/jquery.validate*"));
+						"~/Scripts/jquery-{version}.js",
+                        "~/Scripts/js.cookie.js",
+                            "~/Scripts/jquery.validate*"));
 
 			// Use the development version of Modernizr to develop with and learn from. Then, when you're
 			// ready for production, use the build tool at https://modernizr.com to pick only the tests you need.
@@ -22,44 +18,45 @@ namespace Project
 						"~/Scripts/modernizr-*"));
 
 			bundles.Add(new ScriptBundle("~/bundles/bootstrap").Include(
-					  "~/Scripts/bootstrap.js",
+					  "~/Scripts/bootstrap.min.js",
 					  "~/Scripts/respond.js"));
 
-			bundles.Add(new ScriptBundle("~/bundles/gomc").Include(
-				"~/Scripts/gomc.js"));
+            //bundles.Add(new ScriptBundle("~/bundles/gomc").Include(
+            //	"~/Scripts/gomc.js"));
 
             // Will uncomment once JS is done
-            //bundles.Add(new ScriptBundle("~/bundles/all/js").Include(
-            //    // Remove all blank scripts
-            //    "~/Scripts/gomc.js",
-            //    "~/Scripts/about.js",
-            //    "~/Scripts/admin.js",
-            //    "~/Scripts/docs.js",
-            //    "~/Scripts/downloads.js",
-            //    "~/Scripts/faq.js",
-            //    "~/Scripts/feat.js",
-            //    "~/Scripts/latex-html.js",
-            //    "~/Scripts/latex.js",
-            //    "~/Scripts/login.js",
-            //    "~/Scripts/privacy.js",
-            //    "~/Scripts/publications.js",
-            //    "~/Scripts/site.js",
-            //    "~/Scripts/terms.js",
-            //    "~/Scripts/xml.js",
-            //    "~/Scripts/more-downloads.js",
-            //    "~/Scripts/more-examples.js"
-            //    ));
+            bundles.Add(new ScriptBundle("~/bundles/alljs").Include(
+                // Remove all blank scripts
+                "~/Scripts/gomc-new.js",
+                "~/Scripts/about.js",
+                "~/Scripts/admin.js",
+                "~/Scripts/docs.js",
+                "~/Scripts/downloads.js",
+                "~/Scripts/faq.js",
+                "~/Scripts/feat.js",
+                "~/Scripts/latex-html.js",
+                "~/Scripts/latex.js",
+                "~/Scripts/login.js",
+                "~/Scripts/privacy.js",
+                "~/Scripts/publications.js",
+                "~/Scripts/site.js",
+                "~/Scripts/terms.js",
+                "~/Scripts/xml.js",
+                "~/Scripts/more-downloads.js",
+                "~/Scripts/more-examples.js"
+                ));
 
-            bundles.Add(new StyleBundle("~/Content/css").Include(
-					  "~/Content/bootstrap.css"));
+            bundles.Add(new StyleBundle("~/Content/bootstrap").Include(
+					  "~/Content/bootstrap.css",
+                      "~/Content/bootstrap.theme.css"));
+
+			//bundles.Add(
+			//	new StyleBundle("~/Content/bootstrap").Include(
+			//		"~/Content/bootstrap.css",
+			//		"~/Content/bootstrap.theme.css"));
 
 			bundles.Add(
-				new StyleBundle("~/Content/bootstrap").Include(
-					"~/Content/bootstrap.css",
-					"~/Content/bootstrap.theme.css"));
-
-			bundles.Add(
-				new StyleBundle("~/Content/all/css").Include(
+				new StyleBundle("~/Content/allcss").Include(
                     // Remove all blank files
                     // "~/Content/core.css" // old css
                     "~/Content/gomc.css",
@@ -81,7 +78,7 @@ namespace Project
                     "~/Content/xml.css"
                     ));
             // Set to true to enable full bundle optimization
-            BundleTable.EnableOptimizations = false;
+            BundleTable.EnableOptimizations = true;
         }
 	}
 }
