+++ Project/Views/Home/Gomc.cshtml
@@ -98,7 +98,7 @@
             </div>
         </section>
 </main>
-@section scripts{
+@*@section scripts{
 <script>
         document.onreadystatechange = function () {
             if (document.readyState === 'interactive') {
@@ -131,4 +131,4 @@
             }
         }
 </script>  
-}
+}*@
