+++ Project/Views/Home/_Home_Layout.cshtml
@@ -13,7 +13,7 @@
     <meta charset='utf-8' />
     <meta name="viewport" content="width=device-width, initial-scale=1">
     @Styles.Render("~/Content/bootstrap")
-    @Styles.Render("~/Content/all/css")
+    @Styles.Render("~/Content/allcss")
     @RenderSection("head", false)
 </head>
 <body>
@@ -42,10 +42,7 @@
 </footer>
     @Scripts.Render("~/bundles/jquery")
     @Scripts.Render("~/bundles/bootstrap")
-    @Scripts.Render("~/bundles/js.cookie")
-    @Scripts.Render("~/bundles/jqueryval")
-    @Scripts.Render("~/bundles/gomc")
-<!--The script below is for the google-analytics, as demanded by client, code is pulled from present website-->
+    @Scripts.Render("~/bundles/alljs")
 <script async src="https://www.google-analytics.com/analytics.js"></script>
 <script>
     (function(i, s, o, g, r, a, m) {
